[package] lsof: cross-build properly
I found that the lsof build was using the build host's ar and ranlib
instead of the target's. That's tolerable when the build host is
Linux/ELF. It's not so great when building on Mac OS X.

I also found that lsof was reporting that it was built with the native
compiler instead of the cross-compiler.

Signed-off-by: Mark Mentovai <mark@moxienet.com>